Migrate from map scale to zoom factor
Map scale was a neat, "proper" approach but the back-and-forth between CSS pixels and map units led to a lot of conversion errors that ended up being more hassle than the purity was worth.

The new idea of "zoom" is just the number of map pixels/units in each CSS pixel as a plain factor, which should be a lot easier to work with.